pyproject.toml
[tool.poetry]
name = "projectrag"
version = "0.0.1"
description = "RAG GAMMA"
authors = ["Gajendra"]
readme = "README.md"
packages = [{ include = "*" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

fastapi = ">=0.129.0,<0.130.0"
uvicorn = {extras = ["standard"], version = "^0.41.0"}

sqlalchemy = ">=2.0.46,<3.0.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
pgvector = ">=0.4.2,<0.5.0"

pydantic = { version = ">=2.12.5,<3.0.0", extras = ["email"] }
pyjwt = ">=2.11.0,<3.0.0"
python-jose = ">=3.5.0,<4.0.0"
python-multipart = ">=0.0.22,<0.0.23"

passlib = "1.7.4"
bcrypt = "3.2.2"

pypdf = ">=6.7.0,<7.0.0"
pypdf2 = ">=3.0.1,<4.0.0"
pdfplumber = ">=0.11.9,<0.12.0"

sentence-transformers = ">=5.2.2,<6.0.0"
langchain = ">=1.2.10,<2.0.0"
langchain-community = ">=0.4.1,<0.5.0"
langchain-openai = ">=1.1.9,<2.0.0"
langchain-google-genai = ">=4.2.0,<5.0.0"
google-generativeai = ">=0.8.6,<0.9.0"
openai = ">=2.20.0,<3.0.0"
langsmith = ">=0.7.1,<0.8.0"

torch = { version = "^2.2.0", source = "pytorch-cpu" }
pep8 = "^1.7.1"
pycodestyle = "^2.14.0"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
